+++ src/Symfony/Component/HttpKernel/Bundle/BundleInterface.php
@@ -19,8 +19,6 @@ use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
  * BundleInterface.
  *
  * @author Fabien Potencier <fabien@symfony.com>
- *
- * @method string getPublicDir() Returns relative path for the public assets directory
  */
 interface BundleInterface extends ContainerAwareInterface
 {
@@ -70,4 +68,9 @@ interface BundleInterface extends ContainerAwareInterface
      * @return string The Bundle absolute path
      */
     public function getPath();
+
+    /**
+     * Returns relative path to the public assets directory.
+     */
+    public function getPublicDir(): string;
 }